Rename others submodule references
