+++ python/tdg/docs/other-references/index.rst
@@ -1,4 +1,4 @@
-.. _others:
+.. _references:
 
 Comparing to Other References
 =============================
@@ -18,12 +18,12 @@ Comparing to Other References
 
 We provide some utilities for drawing figures with data from the above sources.
 
-.. autodata:: tdg.others.REFERENCES
+.. autodata:: tdg.references.REFERENCES
    :no-value:
 
-Some references have data that has been superseded by works from the same author, or are far outside the typical parameter range.  Pass a truthy ``include_all`` kwarg to include these in comparison figures. (For an example see :func:`PRL106110403 <tdg.others.PRL106110403.contact_comparison>`.)
+Some references have data that has been superseded by works from the same author, or are far outside the typical parameter range.  Pass a truthy ``include_all`` kwarg to include these in comparison figures. (For an example see :func:`PRL106110403 <tdg.references.PRL106110403.contact_comparison>`.)
 
-.. autofunction:: tdg.others.contact_comparison
+.. autofunction:: tdg.references.contact_comparison
 
-.. autofunction:: tdg.others.energy_comparison
+.. autofunction:: tdg.references.energy_comparison
