+++ python/tdg/tdg/references/CPC2361525.py
@@ -2,7 +2,7 @@
 import torch
 import tdg.conventions
 
-from tdg.others.citation import Citation
+from tdg.references.citation import Citation
 citation = Citation(
     'Krieg et al., Comput. Phys. Commun 236:15-25 (2019)',
     'Krieg:2018pqh'
